SafeDataTool_Backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.22.0
sqlmodel>=0.0.14
sqlalchemy>=2.0.23
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
python-multipart>=0.0.7
pandas>=2.0.0
numpy>=1.24.0
jinja2>=3.1.2
weasyprint>=61.0
scipy>=1.11.0
celery>=5.3.0
redis>=5.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0

